+++ src/main/java/com/github/kd_gaming1/packcore/ui/screen/components/ScreenUIComponents.java
@@ -216,22 +216,22 @@ public class ScreenUIComponents {
 
         if (onSelect != null) {
             //? if <= 1.21.8 {
-            /*entry.mouseDown().subscribe((x, y, button) -> {
+            entry.mouseDown().subscribe((x, y, button) -> {
                 if (button == 0) {
                     onSelect.run();
                     return true;
                 }
                 return false;
             });
-            *///?} else {
-            entry.mouseDown().subscribe((click, doubled) -> {
+            //?} else {
+            /*entry.mouseDown().subscribe((click, doubled) -> {
                 if (click.button() == 0) {
                     onSelect.run();
                     return true;
                 }
                 return false;
             });
-            //?}
+            *///?}
             entry.cursorStyle(CursorStyle.HAND);
         }
     }