+++ src/main/java/com/github/kd_gaming1/packcore/ui/screen/components/WizardUIComponents.java
@@ -100,9 +100,9 @@ public class WizardUIComponents {
 
         if (onClick != null) {
             //? if <= 1.21.8 {
-            /*card.mouseDown().subscribe((x, y, button) -> {
-             *///?} else
-            card.mouseDown().subscribe((click, doubled) -> {
+            card.mouseDown().subscribe((x, y, button) -> {
+             //?} else
+            //card.mouseDown().subscribe((click, doubled) -> {
                 onClick.accept(card);
                 return true;
             });
@@ -287,9 +287,9 @@ public class WizardUIComponents {
 
         if (onClick != null) {
             //? if <= 1.21.8 {
-            /*imageContainer.mouseDown().subscribe((x, y, button) -> {
-             *///?} else
-            imageContainer.mouseDown().subscribe((click, doubled) -> {
+            imageContainer.mouseDown().subscribe((x, y, button) -> {
+             //?} else
+            //imageContainer.mouseDown().subscribe((click, doubled) -> {
                 onClick.run();
                 return true;
             });