+++ src/main/java/com/github/kd_gaming1/packcore/ui/screen/configmanager/SelectiveFileApplicationScreen.java
@@ -22,8 +22,8 @@ import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.gui.screen.Screen;
 import net.minecraft.text.Style;
 //? if >= 1.21.10 {
-import net.minecraft.text.StyleSpriteSource;
-//?}
+/*import net.minecraft.text.StyleSpriteSource;
+*///?}
 import net.minecraft.text.Text;
 import net.minecraft.util.Identifier;
 import org.slf4j.Logger;
@@ -104,9 +104,9 @@ public class SelectiveFileApplicationScreen extends BasePackCoreScreen {
         return Components.label(
                 Text.literal(mode.getTitle() + " - " + sourceName)
                         //? if <= 1.21.8 {
-                        /*.styled(s -> s.withFont(Identifier.of(MOD_ID, "gallaeciaforte")))
-                         *///?} else
-                        .styled(s -> s.withFont(new StyleSpriteSource.Font(Identifier.of(MOD_ID, "gallaeciaforte"))))
+                        .styled(s -> s.withFont(Identifier.of(MOD_ID, "gallaeciaforte")))
+                         //?} else
+                        //.styled(s -> s.withFont(new StyleSpriteSource.Font(Identifier.of(MOD_ID, "gallaeciaforte"))))
         ).color(color(TEXT_PRIMARY));
     }