+++ src/main/java/com/github/kd_gaming1/packcore/ui/screen/scamshield/ScamEducationScreen.java
@@ -12,8 +12,8 @@ import io.wispforest.owo.ui.core.*;
 import net.minecraft.client.gui.screen.Screen;
 import net.minecraft.text.Style;
 //? if >= 1.21.10 {
-import net.minecraft.text.StyleSpriteSource;
-//?}
+/*import net.minecraft.text.StyleSpriteSource;
+*///?}
 import net.minecraft.text.Text;
 import net.minecraft.util.Identifier;
 
@@ -51,9 +51,9 @@ public class ScamEducationScreen extends BasePackCoreScreen {
         return Components.label(
                         Text.literal("Scam Protection Education")
                                 //? if <= 1.21.8 {
-                                /*.styled(s -> s.withFont(Identifier.of(MOD_ID, "gallaeciaforte")))
-                                 *///?} else
-                                .styled(s -> s.withFont(new StyleSpriteSource.Font(Identifier.of(MOD_ID, "gallaeciaforte"))))
+                                .styled(s -> s.withFont(Identifier.of(MOD_ID, "gallaeciaforte")))
+                                 //?} else
+                                //.styled(s -> s.withFont(new StyleSpriteSource.Font(Identifier.of(MOD_ID, "gallaeciaforte"))))
                 ).color(Color.ofRgb(ACCENT_SECONDARY))
                 .shadow(true)
                 .margins(Insets.of(0, 0, 4, 4));