+++ src/main/java/com/github/kd_gaming1/packcore/ui/screen/wizard/BaseWizardPage.java
@@ -15,8 +15,8 @@ import io.wispforest.owo.ui.core.*;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.text.Style;
 //? if >= 1.21.10 {
-import net.minecraft.text.StyleSpriteSource;
-//?}
+/*import net.minecraft.text.StyleSpriteSource;
+*///?}
 import net.minecraft.text.Text;
 import net.minecraft.util.Identifier;
 import net.minecraft.util.Util;
@@ -103,9 +103,9 @@ public abstract class BaseWizardPage extends BaseOwoScreen<StackLayout> {
 
         LabelComponent titleLabel = (LabelComponent) Components.label(pageInfo.title.copy()
                 //? if <= 1.21.8 {
-                /*.styled(s -> s.withFont(Identifier.of(MOD_ID, "gallaeciaforte"))))
-                 *///?} else
-                .styled(s -> s.withFont(new StyleSpriteSource.Font(Identifier.of(MOD_ID, "gallaeciaforte")))))
+                .styled(s -> s.withFont(Identifier.of(MOD_ID, "gallaeciaforte"))))
+                 //?} else
+                //.styled(s -> s.withFont(new StyleSpriteSource.Font(Identifier.of(MOD_ID, "gallaeciaforte")))))
                 .color(Color.ofRgb(ACCENT_SECONDARY))
                 .shadow(true)
                 .margins(Insets.of(0, 0, 4, 4));
@@ -118,9 +118,9 @@ public abstract class BaseWizardPage extends BaseOwoScreen<StackLayout> {
 
         progressIndicator.child(Components.label(Text.literal("| Step " + pageInfo.currentStep() + " of " + pageInfo.totalSteps())
                 //? if <= 1.21.8 {
-                /*.styled(s -> s.withFont(Identifier.of(MOD_ID, "gallaeciaforte"))))
-                 *///?} else
-                .styled(s -> s.withFont(new StyleSpriteSource.Font(Identifier.of(MOD_ID, "gallaeciaforte")))))
+                .styled(s -> s.withFont(Identifier.of(MOD_ID, "gallaeciaforte"))))
+                 //?} else
+                //.styled(s -> s.withFont(new StyleSpriteSource.Font(Identifier.of(MOD_ID, "gallaeciaforte")))))
                 .color(Color.ofRgb(ACCENT_SECONDARY))
                 .shadow(true)
         );