+++ src/main/java/com/github/kd_gaming1/packcore/ui/screen/wizard/pages/ItemBackgroundWizardPage.java
@@ -169,9 +169,9 @@ public class ItemBackgroundWizardPage extends BaseWizardPage {
 
         // Click handler
         //? if <= 1.21.8 {
-        /*imageContainer.mouseDown().subscribe((mouseX, mouseY, button) -> {
-         *///?} else
-        imageContainer.mouseDown().subscribe((click, doubled) -> {
+        imageContainer.mouseDown().subscribe((mouseX, mouseY, button) -> {
+         //?} else
+        //imageContainer.mouseDown().subscribe((click, doubled) -> {
             selectBackground(option.key);
             return true;
         });